si le dossier tmp et xmls existent déjà on les supprime pour ne pas avoir des warning au début

+++ parsers/Monaco/TribunalSupreme/update.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
-
-rm -r tmp
-rm -r xmls
+if [ -d "tmp" ];then
+  rm -r tmp
+fi
 touch all_urls.txt
-mkdir xmls
 mkdir -p tmp/pages
 mkdir -p tmp/home_pages
+if [ -d "xmls" ];then
+  rm -r xmls
+fi
+
+mkdir xmls
+
 cd tmp/
 touch urls.txt
 cd ..
@@ -13,9 +18,7 @@ php homes_downloader.php
 php url_listing.php
 php pages_downloader.php
 nbr_pages=$(ls -A "tmp/pages/" | wc -l)
-echo $nbr_page
 if [ $nbr_pages != '0' ];then
-
   for page in tmp/pages/* ; do
     php parser_htmltoxml.php $page
   done
